p29.27 refactor: thread ids.SubsidiaryID everywhere
Convert the subsidiary entity id from bare int64 to the distinct ids.SubsidiaryID
type across every layer: sqlc models + subsidiary/account/fund/budget/expense/
import/transaction queries, the store seam (tree/CRUD, account<->subsidiary map,
fund/program subsidiary sets, report Scope, budget-plan + expense-report + import
subsidiary), reports drill Scope, synth/fixture builders, the ledgerimport CLI,
the web layer (scope selectors, editor/expense/budget helpers, settings default),
and the full test fan-out.

Purely types, zero behavior change: goldens stay byte-identical. Boundary
discipline (rule 15) — bare int64 only at the SQL scan and HTTP parse seams;
subsidiary-scope sqlc params (balances/activity/native-totals/txn-count) stay
int64 and cast at the store call, while entity queries and FK columns are typed
(the deliberate asymmetry); the scs session seam and render fields stay int64.

+++ internal/reports/account_ledger.go
@@ -117,7 +117,7 @@ func runAccountLedger(ctx context.Context, tk *Toolkit, p Params) (Table, error)
 		// transactions that produced it (as-of, this account+currency, this scope).
 		open := openByCcy[ccy]
 		openDrill := &Drill{
-			Scope:      int64(p.Scope),
+			Scope:      p.Scope,
 			AccountIDs: []int64{int64(p.Account)},
 			Currency:   ccy,
 			Mode:       DrillAsOf,
@@ -138,7 +138,7 @@ func runAccountLedger(ctx context.Context, tk *Toolkit, p Params) (Table, error)
 		// scope (descendant closure), ordered (date, split_id) — the SAME filter the
 		// opening/closing balances close over, so opening + Σ == closing.
 		lines, err := tk.Store().DrillSplits(ctx, store.DrillFilter{
-			Scope:     int64(p.Scope),
+			Scope:     p.Scope,
 			AccountID: int64(p.Account),
 			Currency:  ccy,
 			From:      p.From,
@@ -166,7 +166,7 @@ func runAccountLedger(ctx context.Context, tk *Toolkit, p Params) (Table, error)
 		// Closing row: the cumulative balance as of To. It EQUALS opening + Σlines
 		// (running) by construction; it is drillable to its own as-of transactions.
 		closeDrill := &Drill{
-			Scope:      int64(p.Scope),
+			Scope:      p.Scope,
 			AccountIDs: []int64{int64(p.Account)},
 			Currency:   ccy,
 			Mode:       DrillAsOf,
