+++ internal/reports/form_990.go
@@ -188,7 +188,7 @@ func (b *f990Builder) periodLineDrill(accts []int64, ccys map[string]bool) *Dril
 		ccy = c
 	}
 	return &Drill{
-		Scope:      int64(b.p.Scope),
+		Scope:      b.p.Scope,
 		AccountIDs: dedupSortInts(accts),
 		Currency:   ccy,
 		Mode:       DrillPeriod,
@@ -259,7 +259,7 @@ func (b *f990Builder) programTypeLines(
 	for _, ccy := range sortedKeys(byCcy) {
 		accts := dedupSortInts(acctsByCcy[ccy])
 		d := &Drill{
-			Scope:      int64(b.p.Scope),
+			Scope:      b.p.Scope,
 			AccountIDs: accts,
 			Currency:   ccy,
 			Mode:       DrillPeriod,
@@ -591,7 +591,7 @@ func (b *f990Builder) balanceAccountLine(ctx context.Context, l bsLine, asOf, ta
 	var d *Drill
 	if l.acctID != 0 && len(l.byCcy) == 1 {
 		for ccy := range l.byCcy {
-			d = &Drill{Scope: int64(b.p.Scope), AccountIDs: []int64{l.acctID}, Currency: ccy, Mode: DrillAsOf, AsOf: asOf}
+			d = &Drill{Scope: b.p.Scope, AccountIDs: []int64{l.acctID}, Currency: ccy, Mode: DrillAsOf, AsOf: asOf}
 		}
 	}
 	b.lineRowText(l.name, target, conv, d, 2)
@@ -641,7 +641,7 @@ func (b *f990Builder) unmappedBalanceBucket(ctx context.Context, lines []bsLine)
 			conv = cc
 		}
 	}
-	d := &Drill{Scope: int64(b.p.Scope), AccountIDs: dedupSortInts(accts), Currency: ccy, Mode: DrillAsOf, AsOf: b.p.To}
+	d := &Drill{Scope: b.p.Scope, AccountIDs: dedupSortInts(accts), Currency: ccy, Mode: DrillAsOf, AsOf: b.p.To}
 	b.unmappedRow(b.target, conv, d)
 }
