+++ internal/reports/functional_expenses.go
@@ -277,7 +277,7 @@ func (b *feBuilder) cellDrill(acct AccountID, cl Class, native []CurAmt) *Drill 
 	}
 	class := string(cl)
 	return &Drill{
-		Scope:      int64(b.p.Scope),
+		Scope:      b.p.Scope,
 		AccountIDs: []int64{int64(acct)},
 		Currency:   native[0].Currency,
 		Class:      &class,
